Reopen the repository if we have old version
Since we have the revision number of the last indexed repository
saved in the search index, we can use that as a fast way of checking
whether we need to repoen the repository (because some other thread
or process made a change). We can't simply reopen the repository every
time, as that would be very slow.

Changes made without reindexing will not appear immediately, but that should
be fine. It actually makes the displayed content more consistent with the
search results.